Refine the Authorization Response a Bit
The idea being that an auth request is pretty well defined, and there's
no reason to repeat work like validating clients, etc across grants.

AND while this library will not support implicit grants, we want to
leave the door open for extensibility as well as handling open id
values, etc.